// src/app/models/v_AppointmentRecords.model.ts
export class ViewAppointmentRecords{
    idAppointmentRecord: number;
    dateTimeInitial: Date;
    dateTimeFinal: Date;
    observation: string;
    idReasonAppointment: number;
    reasonAppointment: string;
    idUser: number;
    firstNameUser: string;
    documenUser: string;
    idPatient: number;
    fullName: string;
    documentPatient: string;
    idStatusAppointment: number;
    statusAppointment: string;
    colorStatusAppointment: string;
    idTypeAgenda: number;
    typeAgenda: string;
    price: number;

}